## Issue being fixed or feature implemented
The functional test feature_llmq_data_recover.py is currently the slowest one

## What was done?
See commits. Coderabbit says:

> 

## How Has This Been Tested?
Run 20 jobs in parallel.

This PR: test run takes 3360s accumulated; 90% jobs are finished in interval [159...179] seconds with median 167 seconds.
`develop`: test run takes 4353 s (accumulated); 90% jobs are finished in interval [207..229] seconds with median 217 seconds.
It is 30% improvement for the worst case scenario; for median and for average.

## Breaking Changes
N/A

## Checklist:
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone